+++ src/features/mandatory/seededDeath/constants.ts
@@ -6,6 +6,10 @@ import {
 export const SEEDED_DEATH_DEBUG = true;
 export const SEEDED_DEATH_DEBUFF_FRAMES = 45 * ISAAC_FRAMES_PER_SECOND;
 
+export const SEEDED_DEATH_TIMER_STARTING_X = 65;
+export const SEEDED_DEATH_TIMER_STARTING_Y = 79;
+export const SEEDED_DEATH_TIMER_SEASON_OFFSET_X = 18;
+
 /** The holding item animation lasts 1.3 seconds, so we round up to 2 seconds to be safe. */
 export const DEVIL_DEAL_BUFFER_FRAMES = 2 * GAME_FRAMES_PER_SECOND;